# docs/openapi.json
{
  "openapi": "3.1.0",
...
  "info": {
    "title": "Firestone REST API",
    "version": "0.1.0",
    "description": "Static contract for the existing Firestone v0.1 REST adapter. The API is HTTP/1.1 over a private Unix-domain socket; it has no TCP listener and no bearer-token mode. Socket ownership and mode 0600 are the authentication boundary. The document lists the 18 methods explicitly authored in the Firestone router. Axum also performs its normal implicit HEAD handling for GET registrations, but those incidental HEAD variants are not separate Firestone operations. Unknown paths and unsupported methods both return the stable 404 ErrorEnvelope. JSON successes and errors use application/json without a charset; action progress defaults to application/x-ndjson; logs use text/plain. Closing a mutation stream stops delivery while the shared action continues to a safe point; closing a followed log stream cancels that open-ended read."
  },
  "servers": [
    {
      "url": "http://firestone",
      "description": "Placeholder HTTP authority used by Unix-socket clients such as curl --unix-socket. No TCP connection is made.",
      "x-firestone-transport": "unix",
      "x-firestone-default-socket": "$XDG_RUNTIME_DIR/firestone/serve.sock",
      "x-firestone-fallback-socket": "/tmp/firestone-<uid>/serve.sock",
      "x-firestone-socket-mode": "0600",
...
    }
  ],
  "security": [],
...
  "paths": {
...
    "/v1/machines/{name}/console/ws": {
      "get": {
        "tags": [
          "Machines"
        ],
        "operationId": "openMachineConsoleWebSocket",
        "summary": "Attach the machine console over a WebSocket",
        "description": "Upgrades to a WebSocket carrying the machine's serial console. The single-client console broker is contacted and its acknowledgement read **before** the upgrade, so a console that is already attached is refused with a normal `409` naming `firestone console <name>`, and no WebSocket is created. A machine that is not running is refused with the standard not-running error.\n\nFraming: binary frames are raw terminal bytes in both directions; text frames are JSON control messages. `{\"resize\":{\"rows\":R,\"cols\":C}}` is accepted and ignored on this route, because the guest owns its serial geometry. Any other text frame is ignored. The server closes with a descriptive close frame: code 1000 and reason `machine stopped` on end of file, or code 1011 and the failure detail on a transport error.\n\nThis is a WebSocket handshake, not a JSON route: a successful request answers `101 Switching Protocols` with `Upgrade: websocket`. The handshake headers (`Connection`, `Upgrade`, `Sec-WebSocket-Key`, `Sec-WebSocket-Version`) are the standard RFC 6455 ones and are not repeated here. Firestone negotiates no subprotocol and no extension.",
        "responses": {
          "101": {
            "description": "Upgraded. The connection now carries WebSocket frames."
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "403": {
            "description": "The loopback session gate refused the upgrade. A protocol upgrade must additionally prove same-origin: `Sec-Fetch-Site: same-origin` or `none`, or an `Origin` in the bound loopback allowlist. A Unix listener has no gate."
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "409": {
            "$ref": "#/components/responses/Error409"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          },
          "503": {
            "$ref": "#/components/responses/Error503"
          },
          "504": {
            "$ref": "#/components/responses/Error504"
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/MachineName"
          }
        ]
      }
    },
    "/v1/machines/{name}/shell/ws": {
      "get": {
        "tags": [
          "Machines"
        ],
        "operationId": "openMachineShellWebSocket",
        "summary": "Open an SSH shell over a WebSocket",
        "description": "Upgrades to a WebSocket carrying one interactive shell. The machine must be `running` before the upgrade. After the upgrade the server allocates a host pseudo-terminal and runs the same OpenSSH argv `firestone shell` execs, with the pseudo-terminal slave as the child's stdin, stdout and stderr; the master is relayed to the WebSocket. One session per connection: closing the WebSocket kills the child's process group and reaps it.\n\nFraming: binary frames are raw terminal bytes in both directions; text frames are JSON control messages. `{\"resize\":{\"rows\":R,\"cols\":C}}` applies `TIOCSWINSZ` to the pseudo-terminal master. Any other text frame is ignored. The server closes with a descriptive close frame: code 1000 and reason `machine stopped` when the shell ends, or code 1011 and the failure detail when the session could not be established.\n\nThis is a WebSocket handshake, not a JSON route: a successful request answers `101 Switching Protocols` with `Upgrade: websocket`. The handshake headers (`Connection`, `Upgrade`, `Sec-WebSocket-Key`, `Sec-WebSocket-Version`) are the standard RFC 6455 ones and are not repeated here. Firestone negotiates no subprotocol and no extension.",
        "responses": {
          "101": {
            "description": "Upgraded. The connection now carries WebSocket frames."
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "403": {
            "description": "The loopback session gate refused the upgrade. A protocol upgrade must additionally prove same-origin: `Sec-Fetch-Site: same-origin` or `none`, or an `Origin` in the bound loopback allowlist. A Unix listener has no gate."
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "409": {
            "$ref": "#/components/responses/Error409"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          },
          "503": {
            "$ref": "#/components/responses/Error503"
          },
          "504": {
            "$ref": "#/components/responses/Error504"
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/MachineName"
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorKind": {
        "type": "string",
        "enum": [
          "generic",
          "usage",
          "invalid_spec",
          "not_found",
          "not_running",
          "conflict",
          "already_exists",
          "already_running",
          "busy",
          "dependency",
          "timeout",
          "checksum",
          "interrupted"
        ]
      },
      "ErrorInfo": {
        "type": "object",
        "properties": {
          "kind": {
            "$ref": "#/components/schemas/ErrorKind"
          },
          "message": {
            "type": "string"
          },
          "hint": {
            "type": "string"
          }
        },
        "additionalProperties": false,
        "required": [
          "kind",
          "message"
        ],
        "description": "hint is omitted when the error has no public remediation hint."
      },
      "ErrorEnvelope": {
        "type": "object",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorInfo"
          }
        },
        "additionalProperties": false,
        "required": [
          "error"
        ]
      },
...
    },
    "parameters": {
      "MachineName": {
        "name": "name",
        "in": "path",
        "required": true,
        "description": "Percent-decoded UTF-8 machine name. Core path validation requires one non-empty Unix path component without control characters.",
        "schema": {
          "type": "string",
          "minLength": 1,
          "pattern": "^(?!\\.{1,2}$)[^/\\u0000-\\u001F\\u007F-\\u009F]+$"
        },
        "example": "dev"
      },
...
    },
    "responses": {
      "Error400": {
        "description": "Bad request: usage or invalid_spec.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            },
            "examples": {
              "default": {
                "$ref": "#/components/examples/Error400"
              }
            }
          }
        },
        "x-firestone-error-kinds": [
          "usage",
          "invalid_spec"
        ]
      },
...
      "Error404": {
        "description": "Resource or route not found: not_found.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            },
            "examples": {
              "default": {
                "$ref": "#/components/examples/Error404"
              }
            }
          }
        },
        "x-firestone-error-kinds": [
          "not_found"
        ]
      },
...
      "Error409": {
        "description": "Conflict: conflict, already_running, or busy.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            },
            "examples": {
              "default": {
                "$ref": "#/components/examples/Error409"
              }
            }
          }
        },
        "x-firestone-error-kinds": [
          "conflict",
          "already_running",
          "busy"
        ]
      },
...
      "Error500": {
        "description": "Internal action failure: generic, not_running, already_exists, or interrupted.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            },
            "examples": {
              "default": {
                "$ref": "#/components/examples/Error500"
              }
            }
          }
        },
        "x-firestone-error-kinds": [
          "generic",
          "not_running",
          "already_exists",
          "interrupted"
        ]
      },
...
      "Error503": {
        "description": "Host dependency unavailable or broken: dependency.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            },
            "examples": {
              "default": {
                "$ref": "#/components/examples/Error503"
              }
            }
          }
        },
        "x-firestone-error-kinds": [
          "dependency"
        ]
      },
...
      "Error504": {
        "description": "Bounded action deadline exceeded: timeout.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            },
            "examples": {
              "default": {
                "$ref": "#/components/examples/Error504"
              }
            }
          }
        },
        "x-firestone-error-kinds": [
          "timeout"
        ]
      },
...
    },
    "examples": {
...
      "Error400": {
        "value": {
          "error": {
            "kind": "usage",
            "message": "request body is not valid for this route",
            "hint": "send one application/json object with only the documented fields"
          }
        }
      },
...
      "Error404": {
        "value": {
          "error": {
            "kind": "not_found",
            "message": "no machine named 'dev'",
            "hint": "firestone ls"
          }
        }
      },
...
      "Error409": {
        "value": {
          "error": {
            "kind": "conflict",
            "message": "another operation is active",
            "hint": "wait for it to finish and retry"
          }
        }
      },
...
      "Error500": {
        "value": {
          "error": {
            "kind": "generic",
            "message": "the action could not be completed",
            "hint": "retry the request; if it fails again, report a Firestone bug"
          }
        }
      },
...
      "Error503": {
        "value": {
          "error": {
            "kind": "dependency",
            "message": "/dev/kvm is unavailable",
            "hint": "run firestone doctor"
          }
        }
      },
...
      "Error504": {
        "value": {
          "error": {
            "kind": "timeout",
            "message": "the action timed out",
            "hint": "retry with a larger bounded timeout"
          }
        }
      },
...
    }
  },
...
}